Domain.lockForDelete()
